Make SBP and SBP-struct object attributes static.
This defines a __slots__ attribute for each SBP message (and non-SBP
struct) object, which fixes memory for a fixed set of attributes
instead of allocating a dynamic __dict__. Reduces memory overhead on
using the libsbp's Python bindings when we're instantiating 1k's of
objects. Anecdotal 7x decrease in heap usage:

Partition of a set of 240058 objects. Total size = 117214528 bytes.
 Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
     0 100000  42 104800000  89 104800000  89 dict of sbp.navigation.MsgBaselineECEF
     1 100000  42  6400000   5 111200000  95 sbp.navigation.MsgBaselineECEF

Partition of a set of 140393 objects. Total size = 21265240 bytes.
 Index  Count   %     Size   % Cumulative  % Kind (class / dict of class)
     0 100000  71 15200000  71  15200000  71 sbp.navigation.MsgBaselineECEF

Closes #95.

/cc @mfine @fnoble