python: Change convention for receivers and senders:
Iterators now provide a tuple(msg, metadata). Metadata is a dictionary
with optional fields 'delta' and 'timestamp'.
Senders now take the message, and metadata provided as keyword args.

Also combines FrameSender and FrameReceiver into a single Framer object.
ReceiveHandler renambed back to Handler, and provides a callable interface
that wraps its source's.  This works better with the existing assumption in 
piksi_tools that messages may be sent back to sources.