Update Java parser API
The architecture of the Java parser is updated to be consistent with the
SnapCam project.  This includes development of the parser which does not
execute its own thread, but instead ingests supplied serial data a byte at
the time, and returns the parsed message once an entire message is
received.