Close serial port fd when piksi disconnects [ESD-1111] (#660)
* Close serial port fd when piksi disconnects

When a piksi is connected to a host via USB cable it will create
3 virtual serial ports which appear on linux systems as /dev/ttyACM*.
When Piksi resets or the USB cable is disconnected the kernel will
remove these device nodes automatically.

When console connects to Piksi via one of these ports is will
detect the disconnect event and raise a message in the log. However
it will not close the file descriptor. The kernel will be unable to
fully remove the device since a FD is still open.

When the Piksi reboots and tries to recreate its virtual ports the
linux kernel will be unable to reuse the same device number as before
so long as the previous console process is still running.

Solution - When console detects the piksi disconnect event it should
close the handle. This allows the kernel to reuse the previous device
node as soon as piksi reboots.

* Use hostname instead of IP address