Add Rust u16 array reader-parser with robust Err handling
Because MSG_GROUP_META introduces a new type (array of u16) that was
not used before in any other sbp message, a dedicated parser must exist
for the bindings to be built.
Because buffering is always by chunks of u8, an array of u16 must be
built from reading pairs of u8. Convention is always LittleEndian
throughout the entire sbp protocol.