Add yaml definition of new MSG_SOLN_META message
This message will be part of the Fuzed Wagon, as per the new
Starling Output Design.

One field in the MSG_SOLN_META is called solution_inputs and is an array.
Each element is effectively a u16, the first u8 identifies which sensor we're talking about
and the other u8 has the sensor-specific contents.
For the purpose of documenting these sensor-specific contents, dummy struct types,
noted <Sensor>InputType are created, though not actually used as type.
They're there, in a sense, to describe what is effectively a Union type,
each variant being a specific sensor.