python: better PEP 508 compliance [INFRA-336] (#894)
Expanding on https://github.com/swift-nav/libsbp/pull/889 -- make the AOT/JIT features fully optional by specifying their requirements as "extra" for the Python package.  This requires that users install the sbp package with either `pip install sbp[aot]` or `pip install sbp[jit]` in order to get the native code accelerated `sbp.jit` package.

This was tested (on Windows) by using the new Windows container to build a wheel for the package, and installing the wheel with the `aot` extra enabled:

    pip install .\sbp-3.4.6-cp37-cp37m-win_amd64.whl[jit]

The purpose of the Windows container was two-fold, to make it easier to build an environment for building the Python wheel package, and to workaround a bug that prevented building, which was a combination of certain NumPy versions and Windows versions after 2004 (see [this forum post](https://developercommunity.visualstudio.com/content/problem/1207405/fmod-after-an-update-to-windows-2004-is-causing-a.html) for more details).

Also, added a `FILE` metavar/argument to the Python `sbp2json` tool which defaults to `sys.stdin`, and make it handle missing "extra" features more gracefully.