Update the set string functions for unterminated & null-terminated
Adds a truncate option that, if set, truncates the input string so that
it can fit in the string object. This has also meant that the internal
functions need to be given the size of the string by the user, so that
this can be used will non null terminated strings.