rust: pin vergen (#1217)
Pin vergen to version 3, which is the last version that doesn't use
the enum-iterator crate, this crate pushes the MSRV to a very recent
Rust 1.62-- this is a pretty silly/heavy requirement just for a small
utility crate. Version 3 of vergen crate gets us the functionality we
need and allows people to build with older versions of Rust.