Generate Kaitai Struct Format Descriptions and run Kaitai Struct Compiler to generate language bindings (#1336)
Kaitai Struct is a declarative language used to describe binary data structures. KSY files containing format specifications for all supported SBP messages can be generated using the kaitai target. The resulting files can be found in the kaitai/ksy directory. This target also runs the Kaitai Struct compiler to generate bindings for python and perl. These bindings (along with test cases) can be found in the kaitai/python and kaitai/perl  subdirectories respectively.

Due to the nature of Kaitai Struct, the new bindings do not currently allow for serialisation (i.e. construction of SBP objects). Serialisation for java and python targets is supported as an experimental feature in the development version of Kaitai Struct (>0.10) so this restriction should hopefully be removed in the future.

It should also be noted that the field names used by the KSY bindings are not 1:1 identical to the native bindings due constraints which Kaitai Struct places on various token types, e.g. it is forbidden to use upper case characters in field names.
---------
Co-authored-by: Jason Mobarak <jason@swift-nav.com>