Fix rust and js tests [AP-949] (#1383)
# Description

@swift-nav/devinfra

Not every CI stage ran on https://github.com/swift-nav/libsbp/pull/1382
so when it merged to master a few things broke. These were the JS test
stage and one of the rust linting stages. The fixes are minor,
correcting a badly written test spec with a duplicate YAML key, creating
exceptions for several test cases in the JS tests, and omitting
generation of rust tests files which don't contain any cases

# API compatibility

Does this change introduce a API compatibility risk?

No

## API compatibility plan

If the above is "Yes", please detail the compatibility (or migration)
plan:

N/A

# JIRA Reference

https://swift-nav.atlassian.net/browse/BOARD-XXXX