Use docker in GA [AP-1085] (#1393)
# Description

@swift-nav/devinfra

Make use of `libsbp-build` docker image in the "check generated
artefacts" stage. This bring consistency to the artefacts generated in
the CI check since the canonical way of generating libsbp is via the
docker image. This PR doesn't actually make any changes but should
resolve some of the issues which have been seen recently when generating
libsbp locally versus in CI.

# API compatibility

Does this change introduce a API compatibility risk?

N/A

## API compatibility plan

If the above is "Yes", please detail the compatibility (or migration)
plan:

N/A

# JIRA Reference

https://swift-nav.atlassian.net/browse/AP-1085