Add MsgMeasurementPoint [AP-653] (#1388)
# Description

@swift-nav/devinfra

Create a new message to output profiling information from measurement
points.

Current version of some product will output measurement point profiling
information in the LinuxCPUState message making use of the `cmdline`
field to serialise various bits of information. Some other bits of
information are placed in to the other fields in LinuxCpuState but since
the fields are already defined the names are misleading.

This PR introduces a dedicated message which contains the same
information as is currently sent but broken out in to fields with the
correct types. This removes the need to serialise/deserialise this
information once products are updated to use this message instead.

# API compatibility

Does this change introduce a API compatibility risk?

No

## API compatibility plan

If the above is "Yes", please detail the compatibility (or migration)
plan:

No

# JIRA Reference

https://swift-nav.atlassian.net/browse/AP-653