Remove legacy C API [AP-1684] (#1404)
# Removal Warning

This PR will be merged on **8/4/2024**. On this date the legacy C API
will become unavailable, it will not be reintroduced. On or shortly
after this date libsbp version v6.0.0 will be tagged and released

# Description

@swift-nav/devinfra



Remove legacy C API including all generated code, examples, templates,
and sbpg module.

Remove all references to "v4" API which were lying around in a handful
of places. There is no such thing as the V4 API any more, there is just
the libsbp API

# API compatibility

Yes

Legacy C API completely removed including:
- Any message types in the form `struct msg_xxx_t`
- Functions relating to sending or processing encoded payloads or frames
- payload and frame callback types
- `#define`s message types - in the form `SBP_MSG_XXX`
- C++ `LegacyState` wrapper
- All files under the `libsbp/legacy` directory
- V4 compatibility headers for modern API in `libsbp/v4/<package.h>`


## API compatibility plan

The modern API was introduce to this repository in Sep 2021. The legacy
API was retained but all symbols were marked as deprecated and compile
time messages were output by GCC and Clang whenever a legacy header was
used

In Dec 2023 the legacy was moved to a more difficult to access place. It
was no longer accessible simply by include `libsbp/sbp.h` All headers
were retained but even more compile time messages were added indicating
that these files would be deleted in the next major version

This PR is the final stage of the deprecation plan. Any users of this
repository should have already converted their code to use the moden
API.

# JIRA Reference

https://swift-nav.atlassian.net/browse/AP-1684