Fix kaitai python version for messages ending in a variable-length string (#1468)
The kaitai python bindings contain an internal helper class called
BufferKaitaiStream, which is intended to be a faster work-alike for
the KaitaiStream class provided by the kaitai python runtime. This is
implemented using an inner class called KaitaiStream.IOBytes, which
emulates the io.BytesIO interface expected by KaitaiStream.
Unfortunately the previous implementation of
BufferKaitaiStruct.BytesIO.read() contained a bug which meant that it
did not correctly handle messages (such as MSG_DGNSS_STATUS and
MSG_PROFILING_THREAD_INFO) which end in a variable-length string. In
this case, KaitaiStream.read_bytes_full() would call
BufferKaitaiStruct.BytesIO.read() without any arguments, with the
expectation that it would return all of the data until the end of the
message.