Update from Ubuntu 18 (gcc-6, clang-6) to Ubuntu 20 (gcc-7, clang-6) [AP-3178] (#1473)
# Description

@swift-nav/algint-team

This PR updates the C stages in the CI from relying on Ubuntu 18 (gcc-6,
clang-6) to Ubuntu 20 (gcc-7, clang-6). Ubuntu 18 is EOL since May 2023.
Recently, we observed issues, e.g. in GitHub Actions due to old versions
of GLIBC in Ubuntu 18 jobs. We decided to not further invest time and
energy in maintaining these old pipelines, because:
- No customer directly relies on gcc-6.
- We are planning to update our codebase to C++17, which requires at
least gcc-8 (or 7 without CTAD).
- If we want to have full coverage for all customer required toolchains,
that is a different topic and an additional effort. Currently, that is
not done for dependency repositories like `libsbp`.
- To still have a certain confidence that gcc versions down to 7 are
supported, this PR updates to Ubuntu 20, which is still maintained and
offers gcc 7 in its apt repositories.

# API compatibility

Does this change introduce a API compatibility risk?

No, there are no changes to the code.

## API compatibility plan

If the above is "Yes", please detail the compatibility (or migration)
plan:

<!-- Provide a short explanation plan here -->

# JIRA Reference

https://swift-nav.atlassian.net/browse/AP-3178