Apply new cargo fmt rules & remove deprecated actions (#1509)
# Description

@swift-nav/algint-team

This PR should make no meaningful change to the codebase contained in
libsbp. All changes to the rust code simply are the result of running
`cargo fmt` & `cargo clippy --fix` with the latest version of stable
rust.

This should make it so the CI passes.

I also changed the github actions using actions-rs as it is deprecated

# API compatibility

Does this change introduce a API compatibility risk?

No

## API compatibility plan

If the above is "Yes", please detail the compatibility (or migration)
plan:

<!-- Provide a short explanation plan here -->

# JIRA Reference

N/A

#no_auto_pr